Create independent export template files.
Hot reloading is impacted in development mode when multiple exports
are found in the main template files and causes a loss of local state.

To resolve this I had to move the template exports to their own files
and important them in the main files.
