+++ src/templates/school-programs.js
@@ -3,56 +3,7 @@ import PropTypes from "prop-types";
 import { Helmet } from "react-helmet";
 import { graphql } from "gatsby";
 import Layout from "../components/Layout";
-import PageBuilder from "../components/PageBuilder";
-
-export const SchoolProgramsTemplate = ({
-  title,
-  titleColor,
-  headingImage,
-  pageBuilder,
-  helmet,
-}) => {
-  const backgroundImage = !!headingImage
-    ? headingImage.publicURL
-    : "/static/602986bab4e3eb9b86d275153b37f58c/43a2d/tcs-header.png";
-  const data = pageBuilder ?? [];
-
-  return (
-    <section className="section">
-      {helmet || ""}
-      <div className="container content">
-        <div
-          className="full-width-image margin-top-0"
-          style={{
-            backgroundImage: `url(${backgroundImage})`,
-            backgroundPosition: "left",
-            width: "100%",
-          }}
-        >
-          <div className="columns">
-            <div className="column">
-              <h1 className="title" style={{ color:titleColor, fontSize: "5em" }}>
-                {title}
-              </h1>
-            </div>
-          </div>
-        </div>
-
-        <div>
-          <PageBuilder data={data} />
-        </div>
-      </div>
-    </section>
-  );
-};
-
-SchoolProgramsTemplate.propTypes = {
-  content: PropTypes.node.isRequired,
-  contentComponent: PropTypes.func,
-  description: PropTypes.string,
-  title: PropTypes.string,
-  helmet: PropTypes.object,
-};
+import { SchoolProgramsTemplate } from "./template_exports/school-programs-template";
 
 const SchoolPrograms = ({ data }) => {
   const { markdownRemark: page } = data;
